+++ client/package.json
@@ -17,7 +17,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.1",
     "react-responsive-carousel": "^3.2.23",
-    "react-router-dom": "^5.3.4",
+    "react-router-dom": "^6.14.2",
     "react-slick": "^0.29.0",
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
-    "@vitejs/plugin-react": "^4.0.3",
+    "@vitejs/plugin-react": "^4.0.1",
     "autoprefixer": "^10.4.14",
     "eslint": "^8.44.0",
     "eslint-plugin-react": "^7.32.2",
