+++ client/src/App.jsx
@@ -1,19 +1,18 @@
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import './App.css';
-
-import NavBar from './Componentes/NavBar/NavBar';
-import Home from './Views/Home/Home';
-
+// import { useState } from "react";
+// //import "./App.css";
+// import { Route } from "react-router-dom";
+// import NavBar from "./Componentes/NavBar/NavBar.jsx";
+// import Landing from "./Views/Landing/Landing.jsx";
+// import Home from "./Views/Home/Home.jsx";
+import AppRouter from "./Routes/AppRouter.jsx";
 function App() {
+  //const [count, setCount] = useState(0);
 
   return (
-    <div className="App">
-      <NavBar />
-      <Switch>
-      <Route path='/Productos/page/:page' component={Home} />
-      </Switch>
+    <div>
+      <AppRouter />
     </div>
-  )
+  );
 }
 
-export default App
+export default App;
