+++ client/src/Componentes/Productos/ContenedorCartas/ContenedorCartas.jsx
@@ -1,63 +1,60 @@
-import { Link } from "react-router-dom"; 
-import React, { useState, useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import React, { useEffect } from "react";
+import s from "./ContenedorCartas.module.css";
 import Cartas from "../Cartas/Cartas";
-import { getAllProductos } from "../../../Redux/Actions/action";
-import s from "./ContenedorCartas.module.css"
-import { useParams } from "react-router-dom";
-import PaginationButtons from "../../Paginado/PaginationButtons";
+import { useState } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { getAllProductos, getFiltros } from "../../../Redux/Actions/action";
+import Pagination from "../../Paginado/Paginado";
+import { Link } from "react-router-dom";
 
 
-function ContenedorCartas() {
-  const losProductos = useSelector((state) => state.productos);
-  console.log(losProductos)
-
+function CardsContainer() {
   const dispatch = useDispatch();
+  const losProductos = useSelector((state) => state.productos);
+  const productosFiltrados = useSelector((state) => state.productosFiltrados);
+  //ACÁ SE GUARDAN LOS RESULTADOS DE LA BÚSQUEDA DE LA SEARCHBAR
+  //SI NO ENCUENTRA NINGÚN RESULTADO, DEVUELVE UN [false, 'No hay resultados para la búsqueda']
+  const searchResults = useSelector((state) => state.searchResults);
+
   useEffect(() => {
     dispatch(getAllProductos());
   }, []);
 
-  
+  useEffect(() => {
+    dispatch(getFiltros("price", "asc"));
+  }, [dispatch]);
+  const [showFiltrados, setShowFiltrados] = useState(false);
+  const [currentPage, setCurrentPage] = useState(1);
+  const [postsPerPage, setPostsPerPage] = useState(9);
 
-  const { page } = useParams();
-  const pageNumber = page ? parseInt(page) : 1;
-  const [currentPage, setCurrentPage] = useState(pageNumber);
-  const itemsPerPage = 8;
-  const totalPages = Math.ceil(losProductos.length / itemsPerPage);
+  const lastPostIndex = currentPage * postsPerPage;
+  const firstPostIndex = lastPostIndex - postsPerPage;
 
-  const visiblePeople = losProductos.slice((currentPage - 1) * itemsPerPage, currentPage * itemsPerPage);
+  // const currentPosts = losProductos.slice(firstPostIndex, lastPostIndex);
 
-
+  const currentPosts = showFiltrados
+    ? productosFiltrados.slice(firstPostIndex, lastPostIndex)
+    : losProductos.slice(firstPostIndex, lastPostIndex);
 
 
   return (
-    <div>
-
-    {losProductos.length > 0 ? (
-        <div className="cards-container">
-
-          {visiblePeople.map((product, index) => {
-            return (
-              <Link key={product.id} to={`/detail/${product.id}`}>
-                <Cartas key={index} item={product} />
-              </Link>
-            );
-          })}
-          
-        </div>
-      ) : (
-        <h2>No hay resultados para la búsqueda</h2>
-      )}
-
-      <PaginationButtons
-        currentPage={currentPage}
-        totalPages={totalPages}
-        setCurrentPage={setCurrentPage}
-      /> 
-
+    <div className={s.fondo}>
+      {currentPosts?.map((item) => (
+        <Link key={item.id} to={`/detail/${item.id}`}>
+          <Cartas item={item} />
+        </Link>
+      ))}
+      <div className={s.paginado}>
+        <Pagination
+          totalPost={losProductos.length}
+          postsPerPage={postsPerPage}
+          setCurrentPage={setCurrentPage}
+          currentPage={currentPage}
+        />
+      </div>
     </div>
   );
   
 }
 
-export default ContenedorCartas;
+export default CardsContainer;
