+++ client/src/Componentes/Productos/Filtros/Filtros.module.css
@@ -1,0 +1,4 @@
+.fondo {
+  width: 18%;
+  background-color: rgb(108, 129, 129);
+}
