+++ client/src/Redux/Actions/action.jsx
@@ -1,13 +1,15 @@
 import axios from "axios";
-export const GET_PRODUCT = "GET_PRODUCT";
-export const CLEAR_DETAIL = "CLEAR_DETAIL";
-export const GET_DESCRIPTION = "GET_DESCRIPTION";
-export const GET_PICTURE = "GET_PICTURE";
-export const GET_ALL_PRODUCTOS = "GET_ALL_PRODUCTOS";
-export const GET_CATEGORIES = "GET_CATEGORIES";
-export const GET_CATEGORY = "GET_CATEGORY";
-export const TODOS_FILTROS = "TODOS_FILTROS";
-export const SEARCH_PRODUCTS = "SEARCH_PRODUCTS";
+import {
+  GET_PRODUCT,
+  CLEAR_DETAIL,
+  GET_DESCRIPTION,
+  GET_ALL_PRODUCTOS,
+  GET_PICTURE,
+  GET_CATEGORIES,
+  GET_CATEGORY,
+  TODOS_FILTROS,
+  SEARCH_PRODUCTS,
+} from "./constantes";
 
 export const getAllProductos = () => {
   return async function (dispatch) {
@@ -62,15 +64,14 @@ export const getPicture = (id) => {
   };
 };
 
-
 export const getCategories = () => {
   return async function (dispatch) {
     const apiData = await axios.get(`/categories`);
     const categories = apiData.data;
-    const cleanArr = categories.map(item => ({id: item.id, name:item.name}));
+    console.log(categories, "categorias");
     dispatch({
       type: GET_CATEGORIES,
-      payload: cleanArr,
+      payload: categories,
     });
   };
 };
@@ -149,12 +150,5 @@ export const searchProducts = (words) => {
   };
 };
 
-export const getSearchAdnFilterProducts = (urlData) => {
- /*  try {
-    const getProducts = await axios.get(`/categories/${id}`);
-  } catch (error) {
-    
-  } */
-};
 
 ////http://localhost:3001/products?name=Hub%20Usb
