+++ client/src/Redux/Actions/constantes.jsx
@@ -1,0 +1,9 @@
+export const GET_PRODUCT = "GET_PRODUCT";
+export const CLEAR_DETAIL = "CLEAR_DETAIL";
+export const GET_DESCRIPTION = "GET_DESCRIPTION";
+export const GET_PICTURE = "GET_PICTURE";
+export const GET_ALL_PRODUCTOS = "GET_ALL_PRODUCTOS";
+export const GET_CATEGORIES = "GET_CATEGORIES";
+export const GET_CATEGORY = "GET_CATEGORY";
+export const TODOS_FILTROS = "TODOS_FILTROS";
+export const SEARCH_PRODUCTS = "SEARCH_PRODUCTS";
