+++ client/src/main.jsx
@@ -1,21 +1,17 @@
+import React from "react";
 import ReactDOM from "react-dom/client";
-import { Provider } from "react-redux";
-import store from "./redux/store/store.js";
-import App from "./App";
+import App from "./App.jsx";
 import "./index.css";
-//import { HashRouter } from "react-router-dom";
+import { Provider } from "react-redux";
+import store from "./Redux/Store.jsx";
 import axios from "axios";
-
-import { BrowserRouter } from "react-router-dom";
 
 axios.defaults.baseURL = "http://localhost:3001"
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <Provider store={store}>
-    <BrowserRouter>
+  <React.StrictMode>
+    <Provider store={store}>
       <App />
-    </BrowserRouter>
-  </Provider>
+    </Provider>
+  </React.StrictMode>
 );
-
-
